backend/requirements.txt
fastapi==0.111.*
uvicorn[standard]
sqlmodel
psycopg[binary]
python-slugify
